+++ packages/adapters/solana/src/client.ts
@@ -13,7 +13,7 @@ import {
   SafeLocalStorageKeys
 } from '@reown/appkit-common'
 
-import { SolConstantsUtil, SolHelpersUtil } from '@reown/appkit-utils/solana'
+import { SolConstantsUtil } from '@reown/appkit-utils/solana'
 import { SolStoreUtil } from './utils/SolanaStoreUtil.js'
 import type { Provider } from '@reown/appkit-utils/solana'
 
@@ -47,6 +47,7 @@ import { W3mFrameProviderSingleton } from '@reown/appkit/auth-provider'
 import { ConstantsUtil, ErrorUtil } from '@reown/appkit-utils'
 import { createSendTransaction } from './utils/createSendTransaction.js'
 import { CoinbaseWalletProvider } from './providers/CoinbaseWalletProvider.js'
+import base58 from 'bs58'
 
 export interface AdapterOptions {
   connectionSettings?: Commitment | ConnectionConfig
@@ -130,10 +131,7 @@ export class SolanaAdapter implements ChainAdapter {
     this.appKit = appKit
     this.options = options
     this.caipNetworks = options.networks
-    this.defaultCaipNetwork = SolHelpersUtil.getChainFromCaip(
-      options.networks,
-      SafeLocalStorage.getItem(SafeLocalStorageKeys.ACTIVE_CAIP_NETWORK_ID)
-    )
+    this.defaultCaipNetwork = options.defaultNetwork
 
     if (!projectId) {
       throw new Error('Solana:construct - projectId is undefined')
@@ -182,7 +180,7 @@ export class SolanaAdapter implements ChainAdapter {
 
         // If it's not the auth provider, we should auto connect the provider
         if (chainNamespace === this.chainNamespace || !isAuthProvider) {
-          this.setProvider(externalProvider)
+          await this.setProvider(externalProvider)
         }
       },
 
@@ -200,7 +198,7 @@ export class SolanaAdapter implements ChainAdapter {
 
         const signature = await provider.signMessage(new TextEncoder().encode(message))
 
-        return new TextDecoder().decode(signature)
+        return base58.encode(signature)
       },
 
       estimateGas: async params => {
@@ -240,6 +238,8 @@ export class SolanaAdapter implements ChainAdapter {
       getCapabilities: async () => await Promise.resolve('0x'),
 
       grantPermissions: async () => await Promise.resolve('0x'),
+
+      revokePermissions: async () => await Promise.resolve('0x'),
 
       sendTransaction: async params => {
         if (params.chainNamespace !== CommonConstantsUtil.CHAIN.SOLANA) {