+++ packages/core/exports/vue.ts
@@ -4,9 +4,10 @@ import { ChainController } from '../src/controllers/ChainController.js'
 import type { CaipNetwork, CaipNetworkId } from '@reown/appkit-common'
 import { ref } from 'valtio/vanilla'
 import { onUnmounted } from 'vue'
+import { ConnectionController } from '../src/controllers/ConnectionController.js'
 
 // -- Hooks ------------------------------------------------------------
-export function useAppKitNetwork(): {
+export function useAppKitNetworkCore(): {
   caipNetwork: CaipNetwork | undefined
   chainId: number | string | undefined
   caipNetworkId: CaipNetworkId | undefined
@@ -55,4 +56,12 @@ export function useAppKitAccount() {
     isConnected: Boolean(state.activeCaipAddress),
     status: state.status
   }
+}
+
+export function useDisconnect() {
+  async function disconnect() {
+    await ConnectionController.disconnect()
+  }
+
+  return { disconnect }
 }